[ELY-494] Remove support for the REALM_LIST property from DigestServerFactory
